% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddModules.R
\name{AddModules}
\alias{AddModules}
\title{AddModules}
\usage{
AddModules(
  vModuleIDs,
  strPackageDirectory = getwd(),
  strModuleDirectory = "",
  strType = "package",
  bTabItem = TRUE
)
}
\arguments{
\item{vModuleIDs}{IDs of the shiny module to be added to the application. The source code for each module should be saved in strModuleDirectory with files named "mod_{strModuleID}UI.R" and mod_{strModuleID}Server.R". If desired, submodules may also be included using the following naming convention "mod_{strModuleID}_XXX.R".}

\item{strPackageDirectory}{The directory where the BaSS shiny app located.}

\item{strModuleDirectory}{Location of the Module to be added ot the app. Defaults to inst/_shared/modules.}

\item{bTabItem}{Wrap the UI in a shinyDashboard::TabItem? default = TRUE}
}
\description{
{ This function adds a module to a shiny app created by BaSS.  }
}
