% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddNewTabToApp.R
\name{AddNewTabToApp}
\alias{AddNewTabToApp}
\title{AddNewTabToApp}
\usage{
AddNewTabToApp(strTabName, vSubMenuItemNames = NULL)
}
\description{
{ This funciton will add a new tab to the Shiny App. This function assumes that the working directory
is the main directory of the Shiny app created by this package.  This function creates a module file for the UI and the Server in the Modules directory
of the app, it alters the ShinyUI.R file to insert the calls to  modules so the tab is added to the UI. }
}
