% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddTab.R
\name{AddTab}
\alias{AddTab}
\title{AddTab}
\usage{
AddTab(strModuleID, strType = "package")
}
\arguments{
\item{strModuleID}{ID for new module}

\item{bType}{Type of application. Valid options are 'standalone' or 'package'.}
}
\description{
{ This function quickly adds a tab to the current app using the default parameters. Use addModules() when additional customization is needed.   }
}
