% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddTestToPkg.R
\name{AddTestToPkg}
\alias{AddTestToPkg}
\title{AddTestToPkg}
\usage{
AddTestToPkg(strTestFileName)
}
\arguments{
\item{strFunctionName}{The name of the function to add.}
}
\description{
{ Description - The function will add a new test function to the calculation package.
  This function is intended to be used when a file in the package was already created but
  a test file was not created.  This fuction will create a file named /tests/testthat/test-strTestFileName.R }
}
\examples{
\dontrun{
    AddTestToPkg( "RunSimulation" )  # This would create a test file at /tests/testthat/test-RunSimulation.R
}
}
