% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildAndAddPkgToLocalCran.R
\name{BuildAndAddPkgToLocalCran}
\alias{BuildAndAddPkgToLocalCran}
\title{BuildAndAddPkgToLocalCran}
\usage{
BuildAndAddPkgToLocalCran(lCranRepoDetails, strLocalCranName = "localCRAN")
}
\description{
{ Description: Function to build a package and copy to the local cran repo. }
}
