% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPkgsFromGitHub.R
\name{BuildPkgsFromGitHub}
\alias{BuildPkgsFromGitHub}
\title{BuildPkgsFromGitHub}
\usage{
BuildPkgsFromGitHub(
  vGitHubRepos,
  vPkgNames,
  vRefs,
  vAuthTokens,
  strCloneDir = NULL,
  strLocalCranName = "",
  strDirectoryForLocalCran = ""
)
}
\description{
{ Description: This function pulls GitHub repositories, builds the packages, installs them
and copies it to a local cran like repository. }
}
\details{
and copies it to a local cran like repository.
}
