% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateAppPackageFromProjectCreation.R
\name{CreateAppPackageFromProjectCreation}
\alias{CreateAppPackageFromProjectCreation}
\title{CreateAppPackageFromProjectCreation}
\usage{
CreateAppPackageFromProjectCreation(path, ...)
}
\arguments{
\item{path}{The path that consists of the directory and last folder which is considered the project folder and project name.}
}
\description{
{ Description: This function is intended to be called from the R Studio Project creation widgit, not a user.   }
}
