% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateAppStandalone.R
\name{CreateAppStandalone}
\alias{CreateAppStandalone}
\title{CreateAppStandalone}
\usage{
CreateAppStandalone(
  strDirectory = getwd(),
  strName = "newApp",
  strDisplayName = "",
  strCalculationLibraryName = "",
  strAuthors = ""
)
}
\arguments{
\item{strDirectory}{The directory where the project folder is created.  If this parameter is left blank then the current working directory will be used.}

\item{strName}{{The folder where the app saved.}}

\item{strDisplayName}{{Display name for the app. strName is used by default}}

\item{strCalculationLibraryName}{{Name of the calculation library (if any)}}

\item{strAuthors}{{Author Names. Blank by default}}
}
\description{
{ This function creates a standalone Shiny app.  }
}
