% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateAppStandaloneFromProjectCreation.R
\name{CreateAppStandaloneFromProjectCreation}
\alias{CreateAppStandaloneFromProjectCreation}
\title{CreateAppStandaloneFromProjectCreation}
\usage{
CreateAppStandaloneFromProjectCreation(path, ...)
}
\description{
{ Description: Add Description }
}
