% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBaSSApp.R
\name{CreateBaSSApp}
\alias{CreateBaSSApp}
\title{CreateBaSSApp}
\usage{
CreateBaSSApp(
  strProjectDirectory = "",
  strProjectName = "MyNewProject",
  strCalculationLibraryName = "MyCalculationLibrary",
  strShinyAppName = "MyShinyApp",
  strShinyAppDisplayName = "My Shiny App",
  bCreateWithExampleTabs = TRUE
)
}
\arguments{
\item{strProjectDirectory}{The directory where the project should be created.  If this parameter is left blank then the current working directory will be used.}

\item{strCalculationLibraryName}{{ The name of the R package that is created to contain the calculations for the applicaiton and is referenced in the Shiny app. A folder is created for the
libarary and the R Studio project file will be named, strCalculationLibraryName.Rproj}}

\item{strShinyAppName}{{The name of the R Shiny project.  The project file will  be called strShinyAppName.Rproj and will load the strCalculationLibraryName library as part of the Shiny
app.   }}

\item{strShinyAppDisplayName}{{The name of the applciation to be displayed on the UI that is created.}}

\item{strPojectName}{{The name of the project folder. If strProjectName  is blank or missing then a then the R package and R shiny app projects are created in the strProjectDirectory.
if strProjectName is provided and is not blank then a folder named  strProjectName is created in the strProjectDirectory directory.}}
}
\description{
{Create an application that consists of a Shiny app and a R library package complete with an example testthat unit test.   The Shiny app will reference
the R library so you should build the R library first before trying to run the Shiny app.  }
}
