% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBaSSAppGadget.R
\name{CreateBaSSAppGadget}
\alias{CreateBaSSAppGadget}
\title{CreateBaSSAppGadget}
\usage{
CreateBaSSAppGadget()
}
\description{
{R Studio Gadget to create an application that consists of a Shiny app and a R library package complete with an example testthat unit test.   The Shiny app will reference
the R library so you should build the R library first before trying to run the Shiny app.  }
}
