% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBaSSRPackage.R
\name{CreateBaSSRPackage}
\alias{CreateBaSSRPackage}
\title{CreateBaSSRPackage}
\usage{
CreateBaSSRPackage(
  strProjectDirectory,
  strPackageName,
  strAuthors = NULL,
  bCreateProjectSubdirectory = TRUE
)
}
\arguments{
\item{strProjectDirectory}{The directory where the project should be created.  If this parameter is left blank then the current working directory will be used.}

\item{strPackageName}{{The name of the package folder. If strPackageName  is blank or missing then the project is created in the strProjectDirectory.
if strPackageName is provided, is not blank and bCreateProjectSubdirectory == TRUE then a folder named  strPackageName is created in the strProjectDirectory directory.}}

\item{bCreateProjectSubdirectory}{{If bCreateProjectSubdirectory then then a subdirectory for the project is created in strProjectDirectory.  }}
}
\description{
{Create an R package complete with an example testthat unit test.   }
}
