% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBaSSShinyAppAsPkg.R
\name{CreateBaSSShinyAppAsPkg}
\alias{CreateBaSSShinyAppAsPkg}
\title{CreateBaSSShinyAppAsPkg}
\usage{
CreateBaSSShinyAppAsPkg(
  strProjectDirectory,
  strShinyAppName,
  strShinyAppDisplayName,
  strCalculationLibraryName,
  strAuthors,
  bCreateProjectSubdirectory = TRUE,
  bCreateWithExampleTabs = TRUE
)
}
\description{
{ This function creates a Shiny app as an R Package.  }
}
