% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBaSSShinyAppAsPkg2.R
\name{CreateBaSSShinyAppAsPkg2}
\alias{CreateBaSSShinyAppAsPkg2}
\title{CreateBaSSShinyAppAsPkg2}
\usage{
CreateBaSSShinyAppAsPkg2(
  strProjectDirectory = getwd(),
  strShinyAppName = "newApp",
  strShinyAppDisplayName = "",
  strAuthors = "",
  bCreateProjectSubdirectory = TRUE,
  modules = c("home")
)
}
\description{
{ This function creates a Shiny app as an R Package.  }
}
