% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModule.R
\name{CreateModule}
\alias{CreateModule}
\title{Create Module}
\usage{
CreateModule(
  strModuleID = "newModule",
  strDestDirectory = paste0(getwd(), "/R"),
  strUITemplate = NULL,
  strServerTemplate = NULL,
  lCustomParameters = list()
)
}
\arguments{
\item{strModuleID}{Module ID}

\item{strDestDirectory}{Directory to save new files}

\item{strUITemplate}{Module UI Template}

\item{strServerTemplate}{Module Server Template}
}
\description{
Create Module
}
