% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePackage.R
\name{CreatePackage}
\alias{CreatePackage}
\title{CreatePackage}
\usage{
CreatePackage(strDirectory = getwd(), strName = "newPackage", strAuthors = "")
}
\arguments{
\item{strDirectory}{The directory where the project should be created.  If this parameter is left blank then the current working directory will be used.}

\item{strName}{{The folder where the app saved.}}

\item{strAuthors}{{Author Names. Blank by default}}
}
\description{
{Create an R package complete with an example testthat unit test.   }
}
