% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePackageFromProjectCreation.R
\name{CreatePackageFromProjectCreation}
\alias{CreatePackageFromProjectCreation}
\title{CreatePackageFromProjectCreation}
\usage{
CreatePackageFromProjectCreation(path, ...)
}
\description{
{ Description: Add Description }
}
