% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flattenDirectory.R
\name{flattenDirectory}
\alias{flattenDirectory}
\title{Flatten a directory}
\usage{
flattenDirectory(inDir = "src", outDir = "R", overwrite = TRUE)
}
\arguments{
\item{inDir}{a directory to flatten from}

\item{overwrite}{overwrite files in outDir?}

\item{ourDir}{a directory to flatten to}
}
\value{
list of filenames added to outDir
}
\description{
All files are copied from inDir to outDir. Nested directories are removed. For example, `src/module/home.R` becomes `R/module_home.R`
}
