% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDirectoryAndName.R
\name{GetDirectoryAndName}
\alias{GetDirectoryAndName}
\title{GetDirectoryAndName}
\usage{
GetDirectoryAndName(path)
}
\description{
{ Function to split the path into the directory name (without the last folder)
              and name which is created from the last folder.   }
}
