% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP.R
\docType{package}
\name{PREP}
\alias{PREP}
\alias{PREP-package}
\title{PREP}
\description{
The PREP package was developed to create Shiny projects based on a tempalte and a R package, with testthat unit testing already setup, where
the Shiny app utiizes the R package.
}
\author{
\strong{Maintainer}: J. Kyle Wathen \email{KyleWathen@gmail.com}

Authors:
\itemize{
  \item Nan Chen \email{nchen4@gmail.com} [contributor]
  \item Jeremy Wildfire \email{jwildfire@gmail.com} [contributor]
}

}
