% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateRepoOption.R
\name{UpdateRepoOption}
\alias{UpdateRepoOption}
\title{UpdateRepoOption}
\usage{
UpdateRepoOption(
  strRepoName = "lcran",
  strLocalCRAN = "",
  strURL = "",
  bLinux = TRUE
)
}
\description{
{ Description: This function modifies options( repos ) to include a new cran repo and is used for listing a local CRAN like repository if needed. }
}
