% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddFunctionToPkg.R
\name{AddFunctionToPkg}
\alias{AddFunctionToPkg}
\title{AddFunctionToPkg}
\usage{
AddFunctionToPkg(strFunctionName, strFunctionDescription = "", strPkgDir = "")
}
\arguments{
\item{strFunctionName}{The name of the function to add.}

\item{strFunctionDescription}{A description to include at the top of the file with the function.}

\item{strPkgDir}{The directory of the package to add the function to.  If this parameter is left blank then the current working directory will be used.}
}
\description{
{ Add a new function named strFunctionName to a package.  This function creates a file names R/strFunctionName.R
and adds a test function in the file tests/testthat/test-strFunctionName.R to be used with the testthat package. }
}
