% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateApp.R
\name{CreateApp}
\alias{CreateApp}
\title{CreateApp}
\usage{
CreateApp(
  strDirectory = getwd(),
  strName = "NewApp",
  strDisplayName = "",
  strAuthors = "",
  bPackage = TRUE
)
}
\arguments{
\item{strDirectory}{{The directory where the project should be created.  If this parameter is left blank then the current working directory will be used.}}

\item{strName}{{The name of the folder where the R Shiny app is created.}}

\item{strDisplayName}{{The display name of the app. }}

\item{strAuthors}{{List of authors}}

\item{bPackage}{{Specifies whether the app should be wrapped in a pacakge. Default True}}
}
\description{
{Create an R Shiny app based on a template to help ease development.  The R Shiny App utilizes modeules which is very helpful for large scale applications.  }
}
