% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateAppPackage.R
\name{CreateAppPackage}
\alias{CreateAppPackage}
\title{CreateAppPackage}
\usage{
CreateAppPackage(
  strProjectDirectory = getwd(),
  strName = "newApp",
  strDisplayName = "",
  strAuthors = "",
  strModules = c("Home", "Feedback", "Options"),
  bBuildPackage = TRUE
)
}
\description{
{ This function creates a Shiny app as an R Package.  }
}
