% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateAppStandalone.R
\name{CreateAppStandalone}
\alias{CreateAppStandalone}
\title{CreateAppStandalone}
\usage{
CreateAppStandalone(
  strProjectDirectory,
  strShinyAppName,
  strShinyAppDisplayName,
  strCalculationLibraryName,
  strAuthors,
  bCreateProjectSubdirectory = TRUE,
  bCreateWithExampleTabs = TRUE
)
}
\arguments{
\item{strProjectDirectory}{The directory where the project should be created.  If this parameter is left blank then the current working directory will be used.}

\item{strShinyAppName}{{The name of the folder where the R Shiny app is created. If strShinyAppName  is blank or missing then the project is created in the strProjectDirectory.
If strPackageName is provided, is not blank and bCreateProjectSubdirectory == TRUE then a folder named  strShinyAppName is created in the strProjectDirectory directory.}}

\item{bCreateProjectSubdirectory}{{If bCreateProjectSubdirectory then then a subdirectory for the project is created in strProjectDirectory.  }}

\item{bCreateWithExampleTabs}{{By default, the R Shiny app will contain two example tabs (Simulation and Data analyusis), if bCreateWithExampleTabs = FALSE those tabs are removed.}}

\item{bAppInRFolder}{{Create }}
}
\description{
{Create an R Shiny app based on a template to help ease development.  The R Shiny App utilizes modeules which is very helpful for large scale applications.  }
}
