% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateAppStandalone.R
\name{CreateAppStandalone}
\alias{CreateAppStandalone}
\title{CreateAppStandalone}
\usage{
CreateAppStandalone(
  strDirectory = getwd(),
  strName = "newApp",
  strDisplayName = "",
  strAuthors = "",
  vModuleIDs = c("Home", "Simulation", "Feedback", "Options")
)
}
\arguments{
\item{strDirectory}{The directory where the project folder is created.  If this parameter is left blank then the current working directory will be used.}

\item{strName}{{The folder where the app saved.}}

\item{strDisplayName}{{Display name for the app. strName is used by default}}

\item{strAuthors}{{Author Names. Blank by default}}

\item{vModuleIDs}{{list of module IDs to copy.  The function looks in inst/_shared/modules for files named "mod_{moduleID*}" for each value of vModuleIDs provided; matching files are copied to the new app and initialized as shiny modules in app_ui and app_server. See PREP::add_module() for more detail.}}
}
\description{
{ This function creates a standalone Shiny app.  }
}
