% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePackage.R
\name{CreatePackage}
\alias{CreatePackage}
\title{CreatePackage}
\usage{
CreatePackage(
  strDirectory = getwd(),
  strName = "newPackage",
  strAuthors = "",
  bSampleFiles = TRUE
)
}
\arguments{
\item{strDirectory}{{The directory where the project should be created.  If this parameter is left blank then the current working directory will be used.}}

\item{strName}{{The name of the package folder.}}

\item{bSampleFiles}{{copy sample files? Default: TRUE}}
}
\description{
{Create an R package via usethis::create_package. Optionally include a sample functiona and a testthat unit test.   }
}
