% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlattenDirectory.R
\name{FlattenDirectory}
\alias{FlattenDirectory}
\title{Flatten a directory}
\usage{
FlattenDirectory(strInDir = "src", strOutDir = "R", bOverwrite = TRUE)
}
\arguments{
\item{strInDir}{a directory to flatten from}

\item{strOutDir}{a directory to flatten to}

\item{bOverwrite}{overwrite files in outDir?}
}
\value{
list of filenames added to outDir
}
\description{
All files are copied from inDir to outDir. Nested directories are removed. For example, `src/module/home.R` becomes `R/module_home.R`
}
