% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlattenFile.R
\name{FlattenFile}
\alias{FlattenFile}
\title{Flatten a file}
\usage{
FlattenFile(
  strPath,
  strInDir = "src",
  strOutDir = "R",
  bCopy = TRUE,
  bOverwrite = TRUE
)
}
\arguments{
\item{strPath}{location of file to flatten}

\item{strInDir}{the directory of the file}

\item{strOutDir}{a directory to flatten to}

\item{bCopy}{should the file be copied?}

\item{bOverwrite}{overwrite existing files?}
}
\value{
list of filenames added to outDir
}
\description{
Renames and copies a file removing all directories in `path`.
}
\examples{
# FlattenPath("module/test1.R")
# copies src/module/test1.R to R/module_test1.R

}
